MicroServicesGateway/UtilitiesService/requirements.txt
fastapi
pandas
numpy
pymongo
python-dotenv